backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
autopep8==2.0.4
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.8.0
docopt==0.6.2
docutils==0.20.1
fastapi==0.105.0
feedparser==6.0.10
FLAML==2.1.1
fonttools==4.46.0
gotrue==2.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.3.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
numpy==1.26.2
openai==1.5.0
packaging==23.2
Pillow==10.1.0
pkginfo==1.9.6
postgrest==0.13.0
pyautogen==0.2.2
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
readme-renderer==42.0
realtime==1.0.2
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
storage3==0.7.0
StrEnum==0.4.15
supafunc==0.3.1
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
twine==4.0.2
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==11.0.3
Werkzeug==3.0.1
zipp==3.17.0
